+++ src/main/java/ru/bmstu/labs/queriestfidf/repository/WordRepository.java
@@ -7,7 +7,7 @@ import ru.bmstu.labs.queriestfidf.model.Word;
 import java.util.Optional;
 
 @Repository
-public interface WordRepository extends JpaRepository<Word, Long> {
+public interface WordRepository extends JpaRepository<Word, Integer> {
 
     Optional<Word> getByValue(String value);
